Revert "getmemsize back to size_t"
This reverts commit 7b667c8f3221ccf33bf755a87dc08cecfad895c4.

+++ src/Simulation/MolflowSimFacet.cpp
@@ -246,7 +246,7 @@ int MolflowSimFacet::GetHitsSize(int nbMoments) const {
     + (sh.anglemapParams.record ? (sh.anglemapParams.GetRecordedDataSize()) : 0);
 }
 
-size_t MolflowSimFacet::GetMemSize() const {
+int MolflowSimFacet::GetMemSize() const {
 
     // Init with base size
     int mem_size = SimulationFacet::GetMemSize();
